build: Add touch target.
This new target allows each package's .install stamp file to be touched
without actually performing a rebuild. Since this target explicitly
works around the inter-package dependency handling it should be used
with care.

However, it can be helpful in situations where a package with a long
list of reverse-dependencies needs to be rebuilt (f.e. with debugging
symbols enabled). In such cases all reverse-dependencies would normally
be rebuilt, which could require a substantial amount of time. If you
absolutely know what you're doing (or if you don't care about the
consequences) you can run this target to avoid the rebuild.