build: Add rules for build/.install target.
There are currently no rules to build the build/.install target, causing
unnecessary work because all prerequisites need to be rechecked. This is
solved by providing a simple stamp rule.

Note that the unnecessary work is really not that much. Some rudimentary
testing showed it to take about 10 ms to recreate build/.install on the
largest platform available (~250 packages).