multimedia/linphone: Make speex support optional.
This commit adds an improved version of an older patch that allows speex
support to be explicitly enabled or disabled at build time. A Kconfig
symbol is also provided to select speex support. This defaults to 'n' as
is usual for optional features, so platforms that require it will have
to be updated to override the default.