uClibc: Enable %m specifier for printf.
Newer versions of core/udev use the %m printf specifier to output the
equivalent of strerror(errno) in a thread-safe manner. If the specifier
is not supported by the printf implementation, udev will output strange
('[truncated] m') messages on error. This commit enables support for
this specifier in all uClibc configurations.