net/lldpd: Allow XML output to be enabled.
lldpctl can use libs/libxml2 to generate XML output of LLDP data. The
build system doesn't cope well with a cross-compilation environment
because it uses the xml2-config script which is inherently broken. A new
patch is added to replace the call to xml2-config by a corresponding
lookup via pkg-config.

Also, since the package needs to be reconfigured after the patch has
been applied, the obsolete ltmain sysroot support and autoreconf patches
can be dropped.

Furthermore, the package's watch file and Makefile are updated to
reflect the new download location. The update to the latest version
(0.5.6 as of this writing) is not included because it requires
significant changes to the power-via-MDI patch.