uClibc: Add two patches to fix path name length issues
The uClibc build system doesn't work well with long path names. When
building out-of-tree, an absolute path to the build directory is used,
which in turn causes a number of issues in the build system. The Kconfig
implementation uses a maximum path length of 127 characters, which is
relatively easy to break when building in PBS (I first came across this
issue when building a uClibc-based toolchain). Furthermore the build
system collects a list of files to link into final static and shared
libraries. Unfortunately the list contains the absolute path to each of
the files, which overflows the system-imposed limit for the shell's
argument and environment list (ARG_MAX).

The former problem can easily be fixed by substituting PATH_MAX + 1 for
the path length limit, while the latter is more difficult to fix.
Ideally the build system should probably try and create convenience
libraries containing a smaller number of files and link that into the
final libraries. However that would be a very intrusive change, so an
alternative was chosen for now: when building the final library, we
change into the build directory and strip the build directory from the
path to each file, which will significantly reduce the length of each
parameter listed on the command-line.