packages: Add toolchain/ppl package
This package provides the Parma Polyhedra Library, which is used by the
toolchain/gcc-core and toolchain/gcc packages. By building this in PBS,
the toolchain is decoupled better from the build host.

This was triggered by an update on an ArchLinux system where suddenly
the build host's PPL package was no longer compatible with the one
against which GCC was built.