packages: Add web/nspluginwrapper package
This package provides an NPAPI plugin viewer and wrapper which is useful
to embed proprietary plugins that are built against different versions
of e.g. Gtk+ or to run 32-bit plugins on a 64-bit machine. It is also
useful to move an NPAPI plugin to a separate (sandbox) process that will
allow the browser to survive plugin crashes or hangs.