multimedia/linphone: Fix out-of-tree builds
The linphone upstream tarball ships the config.h header file. The reason
for why this is stupid is that for out-of-tree builds it can happen that
the top source directory gets added to the compiler's include path, and
therefore including config.h will actually include the one shipped in
the source tarball instead of the one generated by the configure script.

This is a problem when users of the linphone package (e.g. PTIP) do not
build the libs/gtk package and therefore do not want to build the setup
wizard. The result is a failed build because the makefile is correctly
generated not to build the setup wizard, but the code that uses the
wizard isn't aware of that because it included the wrong config.h.

I should add that I wasn't able to find out why exactly the config.h is
distributed in the tarball in the first place, otherwise I could have
sent a corresponding patch upstream.