packages: Add virtual/multimedia/libav package.
This virtual package shall be used to select the implementation of libav.
Since libav is a fork of ffmpeg both packages can provide:
 libavcodec, libavdevice, libavfilter, libavformat, libavutil,
 libpostproc. libswresample and libswscale

To avoid a clash between these packages we provide this virtual package so
select either ffmpeg or libav.

This will also update the packages already using FFMPEG.