build-tools/libxml2: Disable python bindings.
Since we do not build packages which requires the bindings, this is the
easiest way to fix building libxml2 without installed python development
packages installed on the build host.

Never the less, this needs to be fixed properly. And because we do build
python as part of the build-tools, we do have all we need to make it work.