Allow user-specific configuration
PBS now looks for user configuration in the following locations:
/etc/pbs.mk, ~/.pbs.mk and pbs.mk in the PBS source tree. Each of these
files is included by the top-level Makefile and usually contains
variable assignments to control the behaviour of PBS.

ccache support can be enabled by setting the CCACHE variable to ccache.
Another useful variable is ROOTFS_MOUNT_TMPFS which can be used to
install the root filesystem into a temporary (RAM) filesystem prior to
the final image creation. This is useful to speed up the extraction
because of reduced disk I/O but has the disadvantage of requiring super
user permissions for the mount operation.

Now with documentation!