packages: Add virtual/udev package
This package is used to provide a dependency on either the legacy
core/udev package or the udev implementation provided by the
core/systemd package.

Eventually the upstream systemd package is supposed to allow the udev
components to be built separately. If that ever happens it should be
possible to remove this workaround.