packages: Update core/systemd to 186
The core/udev package has been merged in this release. Legacy platforms
can still use the core/udev package. Eventually, the upstream systemd
package is supposed to allow the udev components to be built separately,
in the event of which the core/udev package can be upgraded.

core/systemd now also provides the virtual/udev package.