packages: Update libs/gtk to 2.24.11
Xinerama support is disabled because no platform currently uses it.
Furthermore the libs/gtk package never depended on xorg/lib/xinerama, so
it shouldn't be building support anyway because the correct build order
of the packages wouldn't be guaranteed. This could lead to the package
winding up with Xinerama support in one build but without said support
in other builds.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>