web/webkit-gtk: Use flex/bison from the build-tools
Since flex and bison can now be built by PBS as part of the build-tools,
the web/webkit-gtk build can use them to make sure every build host has
the same version.

WebKit fails to build against the latest version of bison (2.6), so this
commit also adds a patch that fixes this issue.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Tested-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>