kernel/linux: Use depmod from build-tools
In order to avoid issues related to the kmod/module-init-tools of the
build host, this changes the Linux kernel build script to make use of
the depmod utility provided by the build-tools.

One current issue is that distributions are starting to transition to
a merged /usr and move Linux kernel modules from /lib/modules into
/usr/lib/modules. Obviously before the same transition happens in PBS
the depmod utility needs to be built to continue looking for modules
in /lib/modules.

While at it, this commit also changes the initrd and rootfs targets to
use the build-tools provided depmod.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Tested-by: Julian Scheel <julian@jusst.de>