build: Move the local source links to the build directory
Source links are generally used to build a platform using a local
working copy of some packages, typically for development purposes.
However when doing release build these links generally have to be
removed, which is error prone.
To make things simpler and safer move the source links to the build
directory, in this way it doesn't influence other builds from the same
source tree.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>