packages: Add build-tools/libffi package
This package provides a version of the libs/libffi package that can be
used during the build process of other packages. Recent versions of GLib
require it to build and a follow-up patch will add the corresponding
dependency.

A lot of the meta data associated with this package is already available
in the libs/libffi package, so the data can be split off into a separate
file and this package can reuse the information to make sure both
packages are kept in sync with respect to their version.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>
Tested-by: Julian Scheel <julian@jusst.de>