packages: Update web/webkit-gtk to 1.10.0
This version of WebKitGtk+ can again be built in parallel, but only if
silent rules are disabled. Most of the patches can be removed because
they have been applied upstream but a new slew of patches is required
to properly build this version.

Furthermore some improvements are made for configurations that build the
WebKit2 API.

This patch is partially based on work by Julian Scheel.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>
Tested-by: Julian Scheel <julian@jusst.de>