scripts: Do not apply patches when building from source link
If a package is built from a source link, do not try to apply patches
because it will most likely fail. It will certainly fail when trying to
build the package a second time as the patches will already have been
applied in the first place.

The assumption here is that the user will take care of applying the
necessary patches when building from a source link.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>