kernel/linux: Always regenerate the version number
The package makefile obtains the version number by sourcing the
kernel-version file. In order for this to work properly, this file needs
to be regenerated whenever the git repository changes, since the version
number embedded in the Linux kernel image will be obtained from the git
describe command, which in turn will change with each new commit or if
the working tree becomes dirty.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>