packages/kernel/linux: Prevent usage of installkernel script.
If the build host contains a installkernel script the kernel's install target
tries to use it to install the kernel. This is a bad idea as it will try to
access the hosts root file system. To avoid this we now explicitly specify the
INSTALLKERNEL parameter to point to some (hopefully) not existing file.

Signed-off-by: Julian Scheel <julian@jusst.de>
Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>