autoconf: Add build tool libtool as dependency of autoconf
libtool is currently part of the toolchain, however the libtoolize
script contain its own install path. This break down when the toolchain
is moved somewhere else as is the case with external toolchains.
To work around this problem use libtool from the build tool by adding
it as dependency to autoconf.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-by: Thierry Reding <thierry.reding@avionic-design.de>