build: Add distcc support
distcc can be enable by defining DISTCC to the binary to use,
generally distcc. If CCACHE is enabled distcc is enabled via
CCACHE_PREFIX, otherwise it is just set as the CC filter.

If NUM_CPU hasn't been set distcc is queried for the number
of available jobs. For this to work the distcc hosts list must be
properly configured in $DISTCC_DIR/hosts.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-by: Thierry Reding <thierry.reding@avionic-design.de>