packages: Update libs/libnl to 3.2.21
Separate packages are now provided that contain the libnl documentation,
so the watch file needs to be fixed up to not match these new tarballs.
Add a patch from git that makes some symbols public again after they
were inadvertently moved to a private header in the 3.2.21 release.

The package is now autoreconfigured so the libtool sysroot support patch
can be dropped.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Tested-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>