uclibc: Fix build against recent Linux kernels
The linux/sysinfo.h header exported by the kernel has recently been
modified to use explicitly sized types as defined elsewhere in the
kernel. Among other things this makes use of the __kernel_ulong_t and
__kernel_long_t types, which are defined in asm-generic/posix_types.h
which is included by the architecture specific posix_types.h. However,
uClibc ships with a custom version of that header which explicitly
conflicts with the Linux kernel header but fails to define the above
two types.

This commit adds a patch that defines the missing types for ARM. Other
architecture may have similar issues but so far no build failures were
encountered.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>