packages: Update multimedia/gstreamer/plugins/base to 1.0.5
Most patches can be removed as they are either no longer relevant or
have been applied upstream. Unfortunately the gstreamer maintainers
think that it is a good idea to check for X headers and libraries
manually, which causes -L/usr/lib to end up in the final linker command
for libraries and therefore breaks cross-compilation. The cross-compile
patch therefore needs to be updated to prevent that from happening.

While at it the plugin and tool version suffix is modified to use the
package's major and minor version numbers so that future updates will
hopefully not have to update these explicitly.

Finally add some missing license information.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Tested-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>