packages: Update libs/pam to 1.1.6
Update location and watch file to reflect the fact that the upstream
package is now maintained on fedorahosted.org. Add missing licensing
information and install the files in /usr instead of /. A new patch is
added that updates configure.in to match the expectations of recent
autotools and fix cross-compilation by adding a patch that defines
BUILD_CPPFLAGS to avoid passing the host CPPFLAGS when building tools
that are run on the build host.

Since the above patches require the package to be autoreconfigured, the
libtool sysroot support patch can be dropped.

This new upstream version also requires libtirpc so a dependency on the
libs/libtirpc package is added.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Tested-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>