packages: Update web/webkit-gtk to 1.10.2
Add a missing dependency on build-tools/ruby which is now required
during the build process. Update patches to apply cleanly and fix the
URL in the watch file. A new patch is also added that fixes build
failures when using the GLib-based Unicode backend. On ARM, the Just-
in-Time (JIT) compiler can safely be enabled now that all remaining
issues seem to have been fixed.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Tested-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>