toolchain: Update GCC to 4.7.2
This update includes a major rework of the devel/gcc-libs package as
required by the GCC update.

Obtain the package version from the compiler instead of the toolchain
GCC package. Doing so allows an external toolchain (such as those built
by pbs-stage1) to be used instead of one built by PBS.

Copying the files is now done in a more target-oriented way instead of
listing all commands explicitly in the .do-install target. Furthermore,
if the toolchain has stack-smash-protection (SSP) support, the
corresponding libraries are also provided by the gcc-libs package. In
addition, header files distributed as part of the standard C++ library
are also provided by the gcc-libs-dev package in order to make them
available in the target sysroot. This is required by some external
toolchains that don't look at their own installation directories for C++
include files.

Shared objects are now looked up via the .so symlink, such that the
soname doesn't have to be hardcoded and therefore this package Makefile
should work better across a wide variety of toolchains.

Finally add missing license information.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Tested-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>