packages: Add xorg/app/evemu package
This package provides the evemu utility which can be used to record
information about and events from an input event. The recorded data can
be replayed on another host, creating a virtual device that looks
identical to all components except the Linux kernel. It is thus ideally
suited for input device debugging.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>