build: Fix SHA256 lookup
The grep pattern that looks up a tarball's SHA256 in checksums/sha256
isn't strict enough. Looking for the atk-2.6.0.tar.xz tarball would
match the at-spi2-atk-2.6.0.tar.xz tarball and flag a checksum mismatch.
A quick fix is to prefix the pattern with download/ to make sure that
whole filenames are matched instead of substrings.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>