multimedia/linphone: Drop dependency on multimedia/ortp
The linphone package ships with its own version of the oRTP library and
links against that, so having a second copy is wasteful.

While at it, make sure to depend on libs/libglade if the Gtk+ frontend
is to be built. The current dependency was conditional on the wrong
Kconfig symbol.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>