xorg/lib/x11: Fix build failure
Add a patch to find the keysym definitions in the sysroot instead of the
build host. This fixes a build failure if the xproto package is not
installed on the build host. The patch requires the package to be auto-
reconfigured, so a dependency on xorg/util/util-macros is required.

Furthermore, set CPPFLAGS_FOR_BUILD so that the makekeys utility can
find the headers provided by xproto in the sysroot.

Originally-by: Sören Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Thierry Reding <thierry.reding@avionic-design.de>
Tested-by: Thierry Reding <thierry.reding@avionic-design.de>
Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>