libs/pulseaudio: Update to 3.0
Because of another issue during configure, we autoreconfigure the package.
While at it, we disable some unused features such as HAL support and manpage
generation.
Also the udev-deprecate patch has been applied upstream so that we can drop
the current patches. But we need to add a new patch for unchecked access,
which can happen when changing the a port.

For arm cpus we need to enable memory-barriers. And finally we add SPEEX
as dependency, to enable pulseaudio's resampling module.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>