packages: Update libs/libffi to 3.0.12
The libffi maintainer apparently though he was being very clever to use
some script to modify LDFLAGS passed to libtool. However that script is
completely broken and breaks sysroot support, so a new patch is needed
to remove this functionality. The patch requires the package to be
autoreconfigured so the libtool sysroot support patch can be dropped.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avonic-design.de>