packages: Fix libdir issue on multiarch hosts.
On fedora 17+ x64 host systems, the configure script selects "lib64"
instead of "lib" as the default libdir. Because the PBS scripts assume
that "lib" is be used, all files installed into e.g. "/usr/lib64" are
not packed and will not be installed into sysroot. So this results in
build errors.
To solve this issue, we need to specify the libdir we actualy use.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Thierry Reding <thierry.reding@avionic-design.de>