libs/ncurses: Disable c++ support for uclibc.
There are a few issues when try to build ncurses with c++ support under
the current uclibc toolchain. Some of the header files use seem to be
outdated and cause build failures.
By disabling the c++ support we can solve this issue and because only a
few packages use c++ code we should be fine for now.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Thierry Reding <thierry.reding@avionic-design.de>