libs/fontconfig: Fix various build issues.
The tests directory can not be build out-of-tree. To solve this issue we add a
new patch to add the correct include path. The second issue is releated to the
documentation generation. Fontconfig creates a tool to generate/modify sgml
files. This tool is build using the cross-compiler. So it can not be executed
on the host. To solve this issue we just disable generating the documentation.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Thierry Reding <thierry.reding@avionic-design.de>