packages: Update libs/harfbuzz from 0.9.12 to 0.9.13.
To avoid autoreconf we pass the CXXFLAGS for configure. This will fix a
linker issue using libraries from the building host.

It might be possible that it would be sufficient to pass only the
--sysroot part, but it is working as it is.
Also it might be worth thinking about to set CXXFLAGS for all packages,
maybe we can skip reconfigure for some other packages as well.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>