busybox: Add a udhcpc patch to fix using -n and -b at the same time
In udhcpc when using -n (exit if lease is not obtained) together with
-b (background if lease is not obtained) and no lease can be obtained
it would background because of -b, try a second time to then exit
because of -n.

Admittedly combining -n and -b is strange, however as -n is part of
the default flags used by ifupdown it can be desirable for -b to work
as expected even if -n has also been given.

This patch simply clear -n from the flags when udhpc goes in the
background, making sure it wont exit if there is another failure.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>