kconfig: Add support for using glob patterns to source files
In some setups it is not desirable to use a fixed list of files,
but glob patterns provide a viable alternative in many cases.
This is provided by a new statement 'glob_source' that take a glob
pattern as parameter. Unlike 'source' that error out if the file
wasn't found, it is not an error if the pattern doesn't match anything.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>