build: Fix package-version and package-name to work with make -n
The package-version and package-name macros rely on the output of
make. However when the top makefiles have been invoked with flags that
change the output of make, like -n, things breaks down.
To fix this override MAKEFLAGS with a fixed version that is
appropriate to invoke the tools.
This fix allow make fs to be run with -n or --print-data-base.

Change-Id: I879aa4c60dcfb0b24b0b22a217ab0b8b9805d65f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/20
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>