libs/pulseaudio: Disable neon support on ARM.
Because the configure does not properly detect NEON support on ARM, we disable
it. Since we currently have no hardware supporting NEON, we can disable it for
all ARM platforms.

As soon as we have some hardware we need to update the arch/arm/Kconfig and
add NEON as regular extention.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>