multimedia/gstreamer/plugins/good: Add patch to make multicast work.
Due to a bug in GIO, the udpsrc does not use a default network interface
for multicast traffic. So using a udp multicast stream with playbin will
not work. This patch sets the default to eth0 which is good enougth for
us right now.
Once GIO has been fixed this can be removed.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>