packages: Update vlc from 2.0.5 to 2.0.6.
Because we have no complete OpenGL implementation on tegra we change
the dependencies a bit. We now select only full gl support if we have
it, same for gles.

Also to avoid autoreconf we pass the CXXFLAGS for configure. It might
be possible that it would be sufficient to pass only the --sysroot but
it is working as it is. It might also be an idea to set CXXFLAGS for
all packages, maybe we can skip reconf for some other packages as well.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>