build: Fix package svn checkout.
When the support for building source distributions was added, the command for
the svn export was overwritten. Because the new commands have also a logical
meaning, we rename original command to avoid the clash. To avoid similar
issues in the future and keep it consistent, we also rename the corresponding
git command.

Change-Id: Iab43074eebfb6b7776c00b5042c756128842d959
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/146
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>