core/e2fsprogs: Fix dependency issue and add help.
With commit 92cf71d we have told e2fsprogs to not build its internal version
of uuid and blkid. But this only works if util-linux is build before. As it
only depends on util-linux when we also use util-linux we don't want to have
a hard dependency. So we add a switch to give the user the control of building
or not building uuid implementation.
While we are doing this we also add a help text for the package and its new
option.

Change-Id: Ie20da8a653e64ead5952650363fe680b72667d35
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/156
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>