libs/pam: Add patch to disable documentation.
When building the documentation, a tool is generated, which depends on flex,
and is used to process the input file for the documentation. This tool can
not be build as it depends on a static library provided by flex. This library
can only been found by the linker, if we set LIBRARY_PATH to the build-tools
library directory. But this might cause new problems, because it could
interfere with the cross-compiler as well. So we decided that it is better
to have a way to completely disable the documentation, as we don't needed
anyway.

This patch is adding a patch for the build-system to add a new configure and
updates the Makefile to use this option. Finally also update the install files.

Change-Id: I1bab44e8526e8136dad301054f5a192bbb50e9d3
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/158
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>