multimedia/libexosip2: Add patch to allow tcp socket reuse.
When we restart remote-control too fast, e.g. due to watchdog reset, and
we run the sip via tcp. It can happens that we can not login anymore,
because the socket is still in use.
This is solved by setting the reuse-address socket option for the sip
socket before binding.

Change-Id: I2f2261a27d92faf247a453c3832473e77db4bc07
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/247
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>