arm: Add tegra3 and neon support.
This adds new options to select the actual silicon implementation of a cpu.
For now we only provide support for NVidia's Tegra2 and Tegra3 SOC.

We need this information to select correct build options and to choose the
correct binary blobs (L4T binaries).

This patch also adds a new build option, depending on Tegra3, to build with
NEON support.

Change-Id: I1445beb91a07b44e8089e3d42909daba7bdc5cee
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/295
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>