build: Distinguish between package and pbs version.
To mark packages that have been built from a local source link pbs extends the
version number of the package with a "+"-sign. As binary hook scripts might rely
on the original package version (ie the linux kernel) pass the plain unchanged
package version to the binary hook scripts additionally. The advantage of this
is that no hacks for stripping pbs version suffixes are required, which may
collide with suffixes in the original package version.

Change-Id: Id120d0c4100502b9720a136d0561af2ebabeacc7
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/460
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>