kernel/linux: Fix build for "+" suffixed versions.
Since commit ce4d57e81e the pbs-binary script provides the plain package version
without pbs specifix suffixes to the binary hook scripts. So we can remove the
hack which cuts of the pbs "+"-suffix and broke kernel build for versions which
had a "+" suffixed by the kernel and not pbs.

Change-Id: I77498eec73f8e320c22997421828a48d15546122
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/461
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>