build: Use the correct package version to extract.
While fixing the kernel build issue, when using a source-link, we introduced a
new issue. Instead of using the new PBS_VERSION, which contains the
source-link suffix, we used the package version to extract the build package.
So we break the build of all other source-link packages, except the kernel.
Using the PBS_VERSION solves this issue.

Change-Id: I9a251454a44adbf1858859c12b5c66ff3e263fc7
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/480
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>