packages: Add devel/file package.
The file command is used to identify a particular file according to the
type of data contained by the file. This can be quite useful when dealing
with binary packages. So it should only be used for development purposes.

This patch adds both the regular package as well as the build-tool
package. This is needed because the package version needs a locally
installed package with the same version.

Change-Id: I11e266252d5b5d3f674697443752620197a1bcca
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/496
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>