devel/glibc: Add patch to allow using make v4.
When using make v4, configure errors telling make 4 is too old since it does
not match the regexp in configure. This adds an patch from archlinux glibc
package to update the regexp checking the make version.

In order to do this, we need to rename the patch dir into patches. Otherwise
no patch will be applied. As a side effect this enables the loader patch,
which should have been applied before already, but was not.

Change-Id: I808c4257a7381a74b5770e5e481724dbc6010105
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>