build: Allow sharing the download directory between PBS instances
Replace the simple wget call with a script, that properly handles
concurrent downloads. A lock file is used to protect each download
and the download is first saved in a temporary file to ensure that
the final file never exists with partial content only.

Additionally the new script shows the download progress to be more
user friendly.

Change-Id: Ib8f9734cfb56824523540fb9ffce00c60683c82f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/814
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>