packages/libs/glib: Fix building with uclibc
When building glib with uclibc the configure script detect libiconv as
being the gnulibc built-in version. However during the build some
header verification fail. To fix just tell the configure script to use
an external libiconv.

With uclibc, libiconv is already in the dependency thru gettext, so no
need to change anything there.

Change-Id: I9b04618115184c1d1216df38ce6ed5853d047bde
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/847
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>