build: Add support for extracting zip files
This is much like all other source extracting rules. However because
it is very common for zip files to miss a top level directory, there
is special support for this case. The target extraction directory is
created if it doesn't exists and the package can redefine the
unzip_pkgtree variable to define where the source are extracted.

So by setting unzip_pkgtree to $(pkgbuildtree) zip files without a top
directory will be properly extracted in a $(PACKAGE)-$(VERSION) sub
directory.

Change-Id: Ib83bb25cda3598a64e5a82ea6aeccd1c43f43415
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/851
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>