skidata/setup: Use the correct platform and source
We need to use the correct kernel sources and config. The main kernel
repository for skidata is github, so use it.
Also we need a kernel config without IRQ support for the adnp driver,
otherwise we can not flash the CPLD because with IRQ support we can not
unload the driver.

Also we need to select the setup platform properly, the old configuration
was generated at a time where the setup was the first platform, which was
changed later to not break the old behavior.

Change-Id: I216649c2d88082c2a149f3abd813c500093e1afa
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/935
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>