build/initrd: Import the PBS config when generating multi.its
Convert the current config to m4 using awk and load it before
generatinf multi.its.m4. This is needed for the Tegra platforms
because T20 and T30 need to use different load address.

We also add the scripts directory to the m4 search path to allow loading
generic scripts from there.

Change-Id: Ic0cdc440b6e17c6a4d677b1575b5ecfc27b6b4c5
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1054
Reviewed-by: Julian Scheel <julian@jusst.de>