build: Add a Kconfig symbol to indicate hard float support
Currently some Makefile use the $(FLOAT) variable as well as some ARM
specific ones to configure floating point support in the packages.
As far as I can tell FLOAT was set by the builtin toolchains, but
since the switch to stage1 it isn't set anymore. Most use of the ARM
specific variables should instead use something more generic.

To solve this we add a new Kconfig variable, ARCH_HAS_HARD_FLOAT, that
should be set by the arch config when an FPU is available. This patch
also update the x86 and ARM config to set this flag.

Change-Id: I63d212ee873cc2795900b382ba11fcadc18b2d80
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1123
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>